+++ dashboard.py
@@ -123,8 +123,8 @@ parser.add_argument('-f', '--file', default='conf/hosts.yaml',
                     help="Path to the config file for the host data, %(default)s by default")
 parser.add_argument('-p', '--port', default=5000,
                     help="Port number to run the web server on, %(default)d by default")
-parser.add_argument('-i', '--interval', default=3,
-                    help="The monitoring system check interval, in minutes. %(default)d by default")
+parser.add_argument('-i', '--interval', default=3, type=int,
+                    help="The host check interval, in minutes. %(default)d by default")
 parser.add_argument('-D', '--debug', action='store_true',
                     help='If the program should run in debug mode')
 
@@ -152,11 +152,11 @@ webAppThread.setDaemon(True)
 webAppThread.start()
 
 logging.info('Starting monitoring check daemon')
-monitor = HostMonitor(args.file)
+monitor = HostMonitor(args.file, args.interval)
 
 while 1:
     logging.debug("Running host check")
     status = monitor.check_hosts()
     utils.write_db(db, utils.HOST_STATUS, status)
 
-    time.sleep(60 * args.interval)
+    time.sleep(60)