+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - added a generic network device, shows up/down status only
+- allow custom override of device icon
 
 ### Changed